Edit header include and include guards for C/C++
Including headers in C is now "inh" and "Inh" which auto-adds a .h
to the included file.

Including headers in C++ is now "inc" and "Inc" and auto-adds a .hpp
to headers surrounded by quotations; for local files.

One can choose an Include guard that has a _H_ or _HPP_